Time: 8 ms (83.25%), Space: 10.1 MB (37.94%) - LeetHub

+++ 215-kth-largest-element-in-an-array/215-kth-largest-element-in-an-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int findKthLargest(vector<int>& nums, int k) {
-        priority_queue<int, vector<int>, greater<int>> pq;
+        priority_queue<short, vector<short>, greater<short>> pq;
         
         for(int &x: nums){
             if(pq.size() < k)
